+++ src/facet/inspection/shap/_shap.py
@@ -14,11 +14,8 @@ from pytools.api import AllTracker
 from pytools.fit import FittableMixin, fitted_only
 from pytools.parallelization import ParallelizableMixin
 
-from facet.inspection._explainer import (
-    BaseExplainer,
-    ExplainerFactory,
-    ParallelExplainer,
-)
+from ...explanation.base import BaseExplainer, ExplainerFactory
+from ...explanation.parallel import ParallelExplainer
 
 log = logging.getLogger(__name__)
 
@@ -85,6 +82,12 @@ class ShapCalculator(
 
     #: The names of the features for which SHAP values were calculated.
     feature_index_: Optional[pd.Index]
+
+    # defined in superclass, repeated here for Sphinx:
+    n_jobs: Optional[int]
+    shared_memory: Optional[bool]
+    pre_dispatch: Optional[Union[str, int]]
+    verbose: Optional[int]
 
     def __init__(
         self,
@@ -197,7 +200,7 @@ class ShapCalculator(
     def shap_values(self) -> pd.DataFrame:
         r"""
         The SHAP values per observation and feature, with shape
-        :math:`(\mathit{n_observations}, \mathit{n_outputs} * \mathit{n_features})`
+        :math:`(n_\mathrm{observations}, n_\mathrm{outputs} * n_\mathrm{features})`
         """
 
         assert self.shap_ is not None, ASSERTION__CALCULATOR_IS_FITTED
@@ -211,8 +214,8 @@ class ShapCalculator(
     def shap_interaction_values(self) -> pd.DataFrame:
         r"""
         The SHAP interaction values per observation and feature pair, with shape
-        :math:`(\mathit{n_observations} * \mathit{n_features},
-                \mathit{n_outputs} * \mathit{n_features})`
+        :math:`(n_\mathrm{observations} * n_\mathrm{features}, n_\mathrm{outputs} *
+        n_\mathrm{features})`
 
         :raise AttributeError: this SHAP calculator does not support interaction values
         """
@@ -228,7 +231,7 @@ class ShapCalculator(
         r"""
         The main effects per observation and featuren (i.e., the diagonals of the
         interaction matrices), with shape
-        :math:`(\mathit{n_observations}, \mathit{n_outputs} * \mathit{n_features})`.
+        :math:`(n_\mathrm{observations}, n_\mathrm{outputs} * n_\mathrm{features})`.
 
         :raise AttributeError: this SHAP calculator does not support interaction values
         """